config/settings/env.py
import os
from dotenv import load_dotenv
from pathlib import Path

BASE_DIR = Path(__file__).resolve().parent.parent.parent
load_dotenv(BASE_DIR / '.env')

def get_env_setting(setting, default=None):
    return os.getenv(setting, default)
