requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.2
django-nested-inline==0.4.6
django-resized==1.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
pytils==0.4.4
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
urllib3==2.5.0

python-dotenv==1.0.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
